Separated models.py into separate classes and adjusted imports on backend
