+++ backend/app.py
@@ -1,9 +1,9 @@
 import os
 from flask import Flask, request, jsonify
 from dotenv import load_dotenv
-from services.user_service import UserService
+# from services.user_service import UserService
 from db import db
-from models import Account
+from models.account import Account
 from routes.courses_route import course_bp
 from routes.planned_courses_route import planned_courses_bp
 from routes.programs_route import programs_bp
@@ -15,12 +15,12 @@ from routes.user_programs_route import users_programs_bp
 from routes.users_route import users_bp
 
 from flask_cors import CORS
-from flask_jwt_extended import (
-    JWTManager,
-    create_access_token,
-    jwt_required,
-    get_jwt_identity,
-)
+#from flask_jwt_extended import (
+#    JWTManager,
+#    create_access_token,
+#    jwt_required,
+#    get_jwt_identity,
+#)
 from datetime import timedelta
 from jwt_helper import init_jwt
