+++ backend/routes/db_planned_courses.py
@@ -1,7 +1,7 @@
 from flask import Blueprint, jsonify, request
-from flask_jwt_extended import jwt_required, get_jwt_identity
+from flask_jwt_extended import jwt_required #, get_jwt_identity
 from services.user_plan_service import (
-    UserPlan,
+    UserPlanService,
 )  # Import the service layer for planned course-related operations
 
 # Define a Blueprint for planned courses
@@ -14,7 +14,7 @@ db_planned_courses_bp = Blueprint("db_planned_courses", __name__)
 def get_planned_courses():
     try:
         # Call the service to get the planned courses for the logged-in user
-        planned_courses = UserPlan.get_planned_courses_for_user()
+        planned_courses = UserPlanService.get_planned_courses_for_user()
 
         if isinstance(
             planned_courses, list
@@ -52,7 +52,7 @@ def add_planned_course():
 
     try:
         # Call the service function to add the course to the user's planned courses
-        response = UserPlan.add_course_to_plan(course_id, plan_id, term, year)
+        response = UserPlanService.add_course_to_plan(course_id, plan_id, term, year)
 
         if isinstance(response, dict):
             return jsonify(response), 201  # Successfully added
@@ -78,7 +78,7 @@ def drop_planned_course():
 
         try:
             # Call the service function to remove the course from the user's planned courses
-            response = UserPlan.drop_course_from_plan(plan_id, course_id)
+            response = UserPlanService.drop_course_from_plan(plan_id, course_id)
 
             if isinstance(response, dict):
                 return jsonify(response), 200  # Successfully removed
