+++ backend/services/course_service.py
@@ -1,20 +1,8 @@
 from db import db
-from datetime import datetime
-from models import (
-    Account,
-    StudentDetails,
-    Course,
-    CourseTaken,
-    Program,
-    StudentProgram,
-    RequirementGroup,
-    DegreePlan,
-    PlannedCourse,
-    SchedulePlan,
-    Section,
-)
+# from datetime import datetime
+from models.course import Course
+from models.course_taken import CourseTaken
 from sqlalchemy.exc import SQLAlchemyError
-
 
 class CourseService:
     # ------------------ COURSE OPERATIONS ------------------
