+++ backend/services/taken_course_service.py
@@ -1,18 +1,7 @@
 from db import db
 from datetime import datetime
-from models import (
-    Account,
-    StudentDetails,
-    Course,
-    CourseTaken,
-    Program,
-    StudentProgram,
-    RequirementGroup,
-    DegreePlan,
-    PlannedCourse,
-    SchedulePlan,
-    Section,
-)
+from models.course import Course
+from models.course_taken import CourseTaken
 from sqlalchemy.exc import SQLAlchemyError
 from services.user_service import UserService
 from services.course_service import CourseService
