+++ backend/services/user_program_service.py
@@ -1,20 +1,10 @@
 from services.program_service import ProgramService
 from services.user_service import UserService
 from db import db
-from datetime import datetime
-from models import (
-    Account,
-    StudentDetails,
-    Course,
-    CourseTaken,
-    Program,
-    StudentProgram,
-    RequirementGroup,
-    DegreePlan,
-    PlannedCourse,
-    SchedulePlan,
-    Section,
-)
+# from datetime import datetime
+from models.student_details import StudentDetails
+from models.program import Program
+from models.student_program import StudentProgram
 from sqlalchemy.exc import SQLAlchemyError
